backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
black==25.1.0
certifi==2025.8.3
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.3
colorama==0.4.6
colorlog==6.9.0
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Django==5.2.1
django-cors-headers==4.7.0
django-mercury-performance==0.0.8
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
Faker==37.3.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
markdown-it-py==4.0.0
mdurl==0.1.2
memory-profiler==0.61.0
msgpack==1.1.1
pillow==11.2.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-decouple==3.8
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
toml==0.10.2
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
zope.interface==7.2
